Move out symbolic derivatives
